PPU/LLVM: Fix unsigned/signed type comparaison warning

+++ rpcs3/Gui/SettingsDialog.cpp
@@ -404,11 +404,11 @@ SettingsDialog::SettingsDialog(wxWindow *parent)
 		txt_dbg_range_min->GetValue().ToLong(&minllvmid);
 		txt_dbg_range_max->GetValue().ToLong(&maxllvmid);
 		Ini.LLVMExclusionRange.SetValue(chbox_core_llvm_exclud->GetValue());
-		Ini.LLVMMinId.SetValue(minllvmid);
-		Ini.LLVMMaxId.SetValue(maxllvmid);
+		Ini.LLVMMinId.SetValue((u32)minllvmid);
+		Ini.LLVMMaxId.SetValue((u32)maxllvmid);
 		long llvmthreshold;
 		txt_llvm_threshold->GetValue().ToLong(&llvmthreshold);
-		Ini.LLVMThreshold.SetValue(llvmthreshold);
+		Ini.LLVMThreshold.SetValue((u32)llvmthreshold);
 		Ini.SPUDecoderMode.SetValue(rbox_spu_decoder->GetSelection());
 		Ini.HookStFunc.SetValue(chbox_core_hook_stfunc->GetValue());
 		Ini.LoadLibLv2.SetValue(chbox_core_load_liblv2->GetValue());
